requirements.txt
python>= 3.10, < 3.12
Django>= 6.0, < 7.0
djangorestframework>= 3.14.0, < 4.0
mysqlclient>= 2.1.0, < 3.0
Pillow>= 10.0.0
drf-nested-routers>= 0.95.0, < 1.0
python-decouple>= 3.8, < 4.0
requests>= 2.31.0, < 3.0
pymemcache == 4.0.0

Note: Enable TokenAuthentication and set default permission classes to IsAuthenticated in DRF settings.
djangorestframework-authtoken>= 1.1.0, < 2.

